refactor: 요구사항에 맞추어 createBulkTodos 구현 위치를 App.tsx로 수정
