+++ src/stores/todoStore.ts
@@ -11,20 +11,8 @@ interface TodoStore {
   setTodos: (newTodos: Todo[]) => void;
 }
 
-function createBulkTodos(): Todo[] {
-  const array: Todo[] = [];
-  for (let i = 1; i < 2500; i++) {
-    array.push({
-      id: i,
-      text: `할 일${i}`,
-      checked: false,
-    });
-  }
-  return array;
-}
-
 const useTodosStore = create<TodoStore>((set) => ({
-  todos: createBulkTodos(),
+  todos: [],
 
   setTodos: (newTodos: Todo[]) => set({ todos: newTodos }),
 }));
